add boilerplate info, docs, organize generators
