+++ .license.ignore
@@ -1,0 +1,3 @@
+./LICENSE
+./Makefile
+./scripts/env_sample
