+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2019, The Tari Developer Community
+Copyright (c) 2022, The Tari Project
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
