+++ rustfmt.toml
@@ -1,0 +1,29 @@
+binop_separator = "Back"
+use_small_heuristics = "default"
+comment_width = 120
+format_strings = true
+group_imports = "StdExternalCrate"
+hard_tabs = false
+imports_layout = "HorizontalVertical"
+imports_granularity = "Crate"
+match_block_trailing_comma = true
+max_width = 120
+newline_style = "Native"
+normalize_comments = true
+reorder_imports = true
+reorder_modules = true
+reorder_impl_items = true
+report_todo = "Never"
+report_fixme = "Never"
+space_after_colon = true
+space_before_colon = false
+struct_lit_single_line = true
+use_field_init_shorthand = true
+use_try_shorthand = true
+unstable_features = true
+format_code_in_doc_comments = true
+where_single_line = true
+wrap_comments = true
+overflow_delimited_expr = true
+edition = "2018"
+ignore=[]
