+++ src/errors.rs
@@ -1,17 +1,22 @@
 // Copyright 2022 The Tari Project
 // SPDX-License-Identifier: BSD-3-Clause
 
-// #![deny(missing_docs)]
+#![deny(missing_docs)]
+
+//! Bulletproof+ error definitions
 
 use thiserror::Error;
 
 /// Represents an error in proof creation, verification, or parsing.
 #[derive(Debug, Error)]
 pub enum ProofError {
+    /// A proof component failed to verify
     #[error("A proof component failed to verify: `{0}")]
     VerificationFailed(String),
+    /// Internal data sizes are inconsistent
     #[error("Internal data sizes are inconsistent: `{0}`")]
     InternalDataInconsistent(String),
+    /// Invalid array/vector length error
     #[error("Invalid array/vector length error: `{0}`")]
     InvalidLength(String),
 }
