+++ src/generators/aggregated_gens_iter.rs
@@ -1,0 +1,40 @@
+// Copyright (c) 2018 Chain, Inc.
+// SPDX-License-Identifier: MIT
+
+// Copyright 2022 The Tari Project
+// SPDX-License-Identifier: BSD-3-Clause
+
+use curve25519_dalek::ristretto::RistrettoPoint;
+
+/// A convenience iterator struct for the generators
+pub struct AggregatedGensIter<'a> {
+    pub(crate) array: &'a Vec<Vec<RistrettoPoint>>,
+    pub(crate) n: usize,
+    pub(crate) m: usize,
+    pub(crate) party_idx: usize,
+    pub(crate) gen_idx: usize,
+}
+
+impl<'a> Iterator for AggregatedGensIter<'a> {
+    type Item = &'a RistrettoPoint;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.gen_idx >= self.n {
+            self.gen_idx = 0;
+            self.party_idx += 1;
+        }
+
+        if self.party_idx >= self.m {
+            None
+        } else {
+            let cur_gen = self.gen_idx;
+            self.gen_idx += 1;
+            Some(&self.array[self.party_idx][cur_gen])
+        }
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        let size = self.n * (self.m - self.party_idx) - self.gen_idx;
+        (size, Some(size))
+    }
+}
