+++ src/range_statement.rs
@@ -1,20 +1,32 @@
 // Copyright 2022 The Tari Project
 // SPDX-License-Identifier: BSD-3-Clause
 
-use crate::errors::ProofError;
-use crate::range_parameters::RangeParameters;
-use curve25519_dalek::ristretto::RistrettoPoint;
-use curve25519_dalek::scalar::Scalar;
+#![deny(missing_docs)]
 
+//! Bulletproof+ generators, vector of commitments, vector of optional minimum promised
+//! values and a vector of optional seed nonces for mask recovery
+
+use curve25519_dalek::{ristretto::RistrettoPoint, scalar::Scalar};
+use zeroize::Zeroize;
+
+use crate::{errors::ProofError, range_parameters::RangeParameters};
+
+/// The range proof statement contains the generators, vector of commitments, vector of optional minimum promised
+/// values and a vector of optional seed nonces for mask recovery
 #[derive(Clone, Debug)]
 pub struct RangeStatement {
+    /// The generators and base points needed for a batch of range proofs
     pub generators: RangeParameters,
-    pub commitments: Vec<RistrettoPoint>, // range statement
+    /// The batch of commitments
+    pub commitments: Vec<RistrettoPoint>,
+    /// Optional minimum promised values
     pub minimum_value_promises: Vec<Option<u64>>,
+    /// Optional seed nonces for mask recovery
     pub seed_nonce: Option<Scalar>,
 }
 
 impl RangeStatement {
+    /// Initialize a new 'RangeStatement' with sanity checks
     pub fn init(
         generators: RangeParameters,
         commitments: Vec<RistrettoPoint>,
@@ -47,5 +59,12 @@ impl RangeStatement {
             minimum_value_promises: minimum_value_promise,
             seed_nonce,
         })
+    }
+}
+
+/// Overwrite secrets with null bytes when they go out of scope.
+impl Drop for RangeStatement {
+    fn drop(&mut self) {
+        self.seed_nonce.zeroize();
     }
 }
